refactor(modal-test): tests are in separate files now and we test clickableSelector and open method
